.history/requirements_20250721172547.txt
EbookLib>=0.18
beautifulsoup4>=4.11.0
lxml>=4.9.0
xhtml2pdf>=0.2.5
python-docx>=0.8.11
markdown>=3.4.0
